pcl: remove redundant use of 'bool' type
The 'typedef unsigned int bool' statement is a compile error with
GCC 15, since c23 defaults 'bool' as a built-in type / reserved
keyword.

The original openssl code from which pcl_vpaes_cbc_encrypt is
copied simply used 'int', the actual method impl is asm code
which just expects an int, and the only callers just pass the
integer value 1. Remove use of the bool typedef since it is
serving no purpose.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>