Add support for building against host CppMicroServices lib
For the aesm_sevice it is desirable to be build against the host
provided CppMicroServices library, such that security updates
automatically propagate from the distro vendor.

This is enabled as an opt-in by setting the make variable

  USE_HOST_CPPMICROSERVICES=1

The in-tree bundle of CppMicroServices ships a 6 year outdated
git snapshot that claims to be version 4.0.0, since that time
upstream reverted the version to 3.x.x. Hence the cmake path
used for the host is looking for 'cppmicroservices3' not
'cppmicroservices4' which is used for the ancient bundled copy.

Signed-off-by: Daniel P. Berrangé <berrange@redhat.com>